+++ packages/amplify-category-geo/src/index.ts
@@ -1,11 +1,11 @@
-import { $TSContext, $TSObject, stateManager } from 'amplify-cli-core';
+import { $TSContext, $TSObject, stateManager, exitOnNextTick, $TSAny } from 'amplify-cli-core';
 import { category } from './constants';
 import * as addCommand from './commands/geo/add';
 import * as updateCommand from './commands/geo/update';
 import * as removeCommand from './commands/geo/remove';
 import * as consoleCommand from './commands/geo/console';
 import * as helpCommand from './commands/geo/help';
-import { getServicePermissionPolicies } from './service-utils/resourceUtils';
+import { getServicePermissionPolicies, verifySupportedRegion, checkAnyGeoResourceExists } from './service-utils/resourceUtils';
 import { ServiceName } from './service-utils/constants';
 import { printer } from 'amplify-prompts';
 
@@ -32,9 +32,19 @@ export const executeAmplifyCommand = async (context: $TSContext) => {
   }
 };
 
-export const handleAmplifyEvent = (context: $TSContext, args: $TSObject) => {
-  printer.info(`${category} handleAmplifyEvent to be implemented`);
-  printer.info(`Received event args ${args}`);
+export const handleAmplifyEvent = async (context: $TSContext, args: $TSAny) => {
+  switch (args.event) {
+    case 'PrePush':
+      if ((await checkAnyGeoResourceExists()) && !verifySupportedRegion()) {
+        const errMessage = 'Failed to create Geo resources';
+        await context.usageData.emitError(new Error(errMessage));
+        printer.info('Remove Geo resources using "amplify remove geo" and retry "amplify push"');
+        exitOnNextTick(1);
+      }
+      break;
+    default:
+      break;
+  }
 };
 
 export const getPermissionPolicies = (context: $TSContext, resourceOpsMapping: $TSObject) => {