+++ packages/amplify-graphql-transformer-interfaces/src/transformer-context/resource-resource-provider.ts
@@ -1,3 +1,4 @@
 export interface TransformerResourceProvider {
   generateResourceName(name: string): string;
+  generateIAMRoleName(name: string): string;
 }